Index

This page lists the core methods and types of the Julia component of the DeepART project.

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

DeepART.DeepART

Methods

Types

Constants

DeepART.DEEPART_VERSION

Docs

This section lists the documentation for every exported name of the DeepART package.

Modules

DeepART.DeepART — Module

A module encapsulating the experimental driver code for the DeepART project.

Imports

The following names are imported by the package as dependencies:

AdaptiveResonance

Base

CUDA

ColorSchemes

Colors

Combinatorics

Core

DataFrames

DataStructures

DelimitedFiles

DocStringExtensions

DrWatson

Flux

ImageShow

Images

JLD2

JSON

MLBase

MLDatasets

MLUtils

NumericalTypeAliases

Parameters

Pkg

Plots

Printf

ProgressMeter

Random

UnicodePlots

Exports

The following names are exported and available when using the package:

DEEPART_VERSION

Functions

Types

Constants

DeepART.DEEPART_VERSION — Constant

DEEPART_VERSION

Description

A constant that contains the version of the installed DeepART.jl package.

This value is computed at compile time, so it may be used to programmatically verify the version of DeepART that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.